improve navigation by initially hiding questions of levels 1-3
This should make it more intuitive to navigate to the page with
exercises by establishing this pattern in the first levels.
For higher levels, this keeps the goal of the level visible.

A new intermediate widget is introduced to help with autoscrolling when
questions are hidden on the first levels.