fix an issue in which level 1 does not immediately unlock
This happened when the cursor is moved back to level 0 before navigating to
the subpage of level 1. The selector for rebuilding questions was too
restrictive, so that the level 1 questions would never rebuild (and thus
not become visible) in this case.

A workaround for this problem in 1.1.2 is to solve level 0 again, or to
change the theme or reset the progress.